+++ tests/__init__.py
@@ -5,8 +5,8 @@
 from .test_django import *
 
 from .test_base import *
-from .test_containers import *
 from .test_declarations import *
+from .test_docs_internals import *
 from .test_faker import *
 from .test_fuzzy import *
 from .test_helpers import *